move calls-specific code in the calls product behind a hook
