+++ app/components/post_list/more_messages/more_messages.tsx
@@ -7,17 +7,11 @@ import Animated, {interpolate, useAnimatedStyle, useSharedValue, withSpring} fro
 import {useSafeAreaInsets} from 'react-native-safe-area-context';
 
 import {resetMessageCount} from '@actions/local/channel';
-import {useCallsState, useChannelsWithCalls, useCurrentCall, useGlobalCallsState, useIncomingCalls} from '@calls/state';
+import {useCallsAdjustment} from '@app/products/calls/hooks';
 import CompassIcon from '@components/compass_icon';
 import FormattedText from '@components/formatted_text';
 import TouchableWithFeedback from '@components/touchable_with_feedback';
 import {Events} from '@constants';
-import {
-    CALL_ERROR_BAR_HEIGHT,
-    CALL_NOTIFICATION_BAR_HEIGHT,
-    CURRENT_CALL_BAR_HEIGHT,
-    JOIN_CALL_BAR_HEIGHT,
-} from '@constants/view';
 import {useServerUrl} from '@context/server';
 import useDidUpdate from '@hooks/did_update';
 import EphemeralStore from '@store/ephemeral_store';
@@ -129,28 +123,7 @@ const MoreMessages = ({
     const underlayColor = useMemo(() => `hsl(${hexToHue(theme.buttonBg)}, 50%, 38%)`, [theme]);
     const styles = getStyleSheet(theme);
     const top = useSharedValue(0);
-
-    // Calls state
-    const incomingCalls = useIncomingCalls().incomingCalls;
-    const channelsWithCalls = useChannelsWithCalls(serverUrl);
-    const callsState = useCallsState(serverUrl);
-    const globalCallsState = useGlobalCallsState();
-    const currentCall = useCurrentCall();
-    const dismissed = Boolean(callsState.calls[channelId]?.dismissed[callsState.myUserId]);
-    const inCurrentCall = currentCall?.id === channelId;
-    const joinCallBannerVisible = Boolean(channelsWithCalls[channelId]) && !dismissed && !inCurrentCall;
-
-    // Do we have calls banners?
-    const currentCallBarVisible = Boolean(currentCall);
-    const micPermissionsError = !globalCallsState.micPermissionsGranted && (currentCall && !currentCall.micPermissionsErrorDismissed);
-    const callQualityAlert = Boolean(currentCall?.callQualityAlert);
-    const incomingCallsShowing = incomingCalls.filter((ic) => ic.channelID !== channelId);
-    const callsIncomingAdjustment = (incomingCallsShowing.length * CALL_NOTIFICATION_BAR_HEIGHT) + (incomingCallsShowing.length * 8);
-    const callsAdjustment = (currentCallBarVisible ? CURRENT_CALL_BAR_HEIGHT + 8 : 0) +
-        (micPermissionsError ? CALL_ERROR_BAR_HEIGHT + 8 : 0) +
-        (callQualityAlert ? CALL_ERROR_BAR_HEIGHT + 8 : 0) +
-        (joinCallBannerVisible ? JOIN_CALL_BAR_HEIGHT + 8 : 0) +
-        callsIncomingAdjustment;
+    const callsAdjustment = useCallsAdjustment(serverUrl, channelId);
 
     // The final top:
     const adjustedTop = insets.top + callsAdjustment;
